convert unit test to integration test
