--- packages/datadog-plugin-express/test/middleware-disabled-test/middleware-disabled-plugin.spec.js
@@ -1,5 +1,0 @@
-'use strict'
-
-const test = require('./middleware-disabled-tests')
-
-test('plugin', [{ middleware: false }, { client: false }], {})
