--- packages/datadog-plugin-express/test/middleware-disabled-test/middleware-disabled-tracer.spec.js
@@ -1,5 +1,0 @@
-'use strict'
-
-const test = require('./middleware-disabled-tests')
-
-test('tracer', [{}, { client: false }], { middleware: false })
